Split CLI parser checks into dedicated script
- Move parser error display and numeric separator tests out of the general CLI suite
- Add CI and PR workflow steps for the new parser-focused check
