+++ scripts/test-cli.ts
@@ -4,7 +4,7 @@
  *
  * Common CLI flags tested across all apps: stdin smoke, --help, --unsafe-ffi,
  * --asi, --compat-var, --mode, --timeout, --max-instructions, --max-memory,
- * --stack-size, error display, numeric separator rejection, --log, example scripts.
+ * --stack-size, --log, example scripts.
  */
 
 import { $ } from "bun";
@@ -272,63 +272,6 @@ console.log("--stack-size (bytecode trampoline)...");
   const src = "let n = 0; const f = () => { n++; if (n < 20000) f(); }; f(); console.log(n);";
   const out = await $`echo ${src} | ${LOADER} --mode=bytecode --stack-size=0`.text();
   if (!out.includes("20000")) throw new Error(`Trampoline should reach 20000, got: ${out}`);
-}
-
-// -- Error display (Loader) -----------------------------------------------------
-
-console.log("Error display (SyntaxError with caret and suggestion)...");
-{
-  const res = await $`echo 'const x = 1\nconst y = x +' | ${LOADER} 2>&1`.nothrow();
-  const out = res.text();
-  if (!out.includes("SyntaxError")) throw new Error(`Expected SyntaxError, got: ${out}`);
-  if (!out.includes("^")) throw new Error(`Expected caret in error display, got: ${out}`);
-  const lower = out.toLowerCase();
-  if (!lower.includes("suggest") && !lower.includes("expect") && !lower.includes("unexpected")) {
-    throw new Error(`Expected suggestion/expectation hint in error display`);
-  }
-}
-
-console.log("Error display (JSON error output)...");
-{
-  const proc = Bun.spawnSync([LOADER, "--output=json"], {
-    stdin: new TextEncoder().encode("const x = ;\n"),
-    stdout: "pipe",
-    stderr: "pipe",
-  });
-  const json = JSON.parse(proc.stdout.toString());
-  if (json.ok !== false) throw new Error(`JSON error ok should be false`);
-  if (json.error?.type !== "SyntaxError") throw new Error(`Expected SyntaxError, got ${json.error?.type}`);
-  if (json.error?.line === undefined) throw new Error(`JSON error should include line`);
-  if (json.error?.column === undefined) throw new Error(`JSON error should include column`);
-}
-
-// -- Numeric separator rejection (9 cases) --------------------------------------
-
-console.log("Numeric separator rejection...");
-{
-  const cases: [string, string][] = [
-    ["1_000_", "trailing underscore"],
-    ["0x_FF", "leading underscore after 0x"],
-    ["0b_10", "leading underscore after 0b"],
-    ["0o_77", "leading underscore after 0o"],
-    ["1__000", "consecutive underscores"],
-    ["0_1", "underscore after leading 0"],
-    ["1._0", "underscore after decimal point"],
-    ["1_.0", "underscore before decimal point"],
-    ["1e_10", "underscore after exponent"],
-  ];
-
-  for (const [literal, desc] of cases) {
-    const proc = Bun.spawnSync([LOADER, "--output=json"], {
-      stdin: new TextEncoder().encode(`${literal};\n`),
-      stdout: "pipe",
-      stderr: "pipe",
-    });
-    const json = JSON.parse(proc.stdout.toString());
-    if (json.ok !== false || json.error?.type !== "SyntaxError") {
-      throw new Error(`Numeric separator "${literal}" (${desc}) should be SyntaxError, got ok=${json.ok} type=${json.error?.type}`);
-    }
-  }
 }
 
 // -- --log flag (Loader) --------------------------------------------------------
